+++ contest/AtCoder/abc023/D.cpp
@@ -5,8 +5,8 @@ int main() {
   cin >> n;
   vector<int> h(n), s(n);
   for (int i = 0; i < n; ++i) cin >> h[i] >> s[i];
-  auto f = [&](long long m){
-    vector<long long> v;
+  auto f = [&](int64_t m){
+    vector<int64_t> v;
     for (int i = 0; i < n; ++i) v.emplace_back((m - h[i]) / s[i]);
     sort(v.begin(), v.end());
     for (int i = 0; i < n; ++i) {
@@ -14,5 +14,5 @@ int main() {
     }
     return true;
   };
-  cout << binarySearch(f, 0, numeric_limits<long long>::max()) << endl;
+  cout << binarySearch(f, 0, numeric_limits<int64_t>::max()) << endl;
 }
