+++ contest/yukicoder/128.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long n, m;
+  int64_t n, m;
   cin >> n >> m;
   cout << n / 1000 / m * 1000 << endl;
 }
