+++ contest/yukicoder/166.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long h, w, n, k;
+  int64_t h, w, n, k;
   cin >> h >> w >> n >> k;
   cout << ((h * w - 1) % n == k - 1 ? "YES" : "NO") << endl;
 }
