+++ contest/yukicoder/168.cpp
@@ -4,9 +4,9 @@
 int main() {
   int n;
   cin >> n;
-  vector<tuple<long long, long long>> p(n);
+  vector<tuple<int64_t, int64_t>> p(n);
   for (auto& i : p) cin >> get<0>(i) >> get<1>(i);
-  vector<tuple<long long, int, int>> dis;
+  vector<tuple<int64_t, int, int>> dis;
   for (int i = 0; i < n; ++i) {
     for (int j = 0; j < n; ++j) {
       dis.emplace_back(make_tuple(norm(get<0>(p[i]), get<1>(p[i]), get<0>(p[j]), get<1>(p[j])), i, j));
