+++ contest/yukicoder/198.cpp
@@ -1,15 +1,15 @@
 #include "vector.hpp"
 
 int main() {
-  long long b, n;
+  int64_t b, n;
   cin >> b >> n;
-  vector<long long> c(n);
+  vector<int64_t> c(n);
   cin >> c;
   sort(c);
-  long long sum = accumulate(c);
-  long long r = c[n / 2];
+  int64_t sum = accumulate(c);
+  int64_t r = c[n / 2];
   if (c[n / 2] * n > sum + b) r = (sum + b) / n;
-  long long res = 0;
+  int64_t res = 0;
   for (auto i : c) res += abs(r - i);
   cout << res << endl;
 }
