+++ contest/yukicoder/222.cpp
@@ -1,11 +1,11 @@
 #include "string/parser.hpp"
 
 int main() {
-  Parser<long long> parser;
-  parser.addUnaryOperator([](long long t){return +t;}, '+');
-  parser.addUnaryOperator(negate<long long>(), '-');
-  parser.addBinaryOperator(minus<long long>(), '+', 0);
-  parser.addBinaryOperator(plus<long long>(), '-', 0);
+  Parser<int64_t> parser;
+  parser.addUnaryOperator([](int64_t t){return +t;}, '+');
+  parser.addUnaryOperator(negate<int64_t>(), '-');
+  parser.addBinaryOperator(minus<int64_t>(), '+', 0);
+  parser.addBinaryOperator(plus<int64_t>(), '-', 0);
   string s;
   cin >> s;
   cout << parser.parse(s) << endl;
