+++ contest/yukicoder/237.cpp
@@ -6,7 +6,7 @@ int main() {
   int f[] = {3, 5, 17, 257, 65537};
   int res = -2;
   for (auto i : PowerSet(5)) {
-    long long k = 1;
+    int64_t k = 1;
     for (auto j : i) k *= f[j];
     res += most_bit(a / k) + 1;
   }
