+++ contest/yukicoder/36.cpp
@@ -1,7 +1,7 @@
 #include "math/prime.hpp"
 
 int main() {
-  long long n;
+  int64_t n;
   cin >> n;
   Prime prime;
   cout << (prime.factor(n).size() >= 3u ? "YES" : "NO") << endl;
