+++ contest/yukicoder/71.cpp
@@ -1,7 +1,7 @@
 #include "template.hpp"
 
 int main() {
-  long long n;
+  int64_t n;
   cin >> n;
   cout << (n / 2 + 1) * ((n + 1) / 2 + 1) - 1 << endl;
 }
